requirements.txt
pandas
torch
torchvision
scikit-learn
matplotlib
numpy
pillow
